src/main/java/com/portal/model/PaymentMethod.java
package com.portal.model;

import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.Null;
import javax.validation.constraints.Size;

import com.portal.dto.PaymentMethodDTO;
import com.portal.validators.NotNullNotZero;
import com.portal.validators.ValidationHelper.OnSave;
import com.portal.validators.ValidationHelper.OnUpdate;

import lombok.Builder;
import lombok.EqualsAndHashCode;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@Builder
@Getter
@Setter
@ToString
@NoArgsConstructor
@EqualsAndHashCode(onlyExplicitlyIncluded = true)
public class PaymentMethod {

	@EqualsAndHashCode.Include
	@Null(groups = {OnSave.class})
	@NotNullNotZero(groups = {OnUpdate.class})
	private Integer id;
	
	@NotBlank(groups = {OnUpdate.class, OnSave.class})
	@Size(max = 45, groups = {OnUpdate.class, OnSave.class})
	private String name;
	
	@NotNull(groups = {OnUpdate.class, OnSave.class})
	private Boolean active;

	public PaymentMethod( Integer id, String name, Boolean active ) {
		this.id = id;
		this.name = name;
		this.active = active;
	}
	
	public static PaymentMethod toEntity( PaymentMethodDTO dto ) {
		
		if( dto == null ) {
			return null;
		}
		
		return PaymentMethod.builder()
						.id( dto.getId() )
						.name( dto.getName() )
						.active( dto.getActive() )
						.build();
	}
}
